requirements.txt
httpx==0.24.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
fastmcp==0.1.2
asyncio==3.4.3
